Moved CoordinationService from Experimental to Core
