+++ Core/nbproject/project.properties
@@ -15,6 +15,10 @@ file.reference.StixLib.jar=release/modules/ext/StixLib.jar
 file.reference.tika-core-1.5.jar=release/modules/ext/tika-core-1.5.jar
 file.reference.Tsk_DataModel_PostgreSQL.jar=release/modules/ext/Tsk_DataModel_PostgreSQL.jar
 file.reference.xmpcore-5.1.2.jar=release/modules/ext/xmpcore-5.1.2.jar
+file.reference.curator-client-2.8.0.jar=release/modules/ext/curator-client-2.8.0.jar
+file.reference.curator-framework-2.8.0.jar=release/modules/ext/curator-framework-2.8.0.jar
+file.reference.curator-recipes-2.8.0.jar=release/modules/ext/curator-recipes-2.8.0.jar
+file.reference.zookeeper-3.4.6.jar=release/modules/ext/zookeeper-3.4.6.jar
 javac.source=1.8
 javac.compilerargs=-Xlint -Xlint:-serial
 javadoc.reference.metadata-extractor-2.8.1.jar=release/modules/ext/metadata-extractor-2.8.1-src.zip
@@ -23,5 +27,6 @@ nbm.homepage=http://www.sleuthkit.org/
 nbm.module.author=Brian Carrier
 nbm.needs.restart=true
 source.reference.metadata-extractor-2.8.1.jar=release/modules/ext/metadata-extractor-2.8.1-src.zip!/Source/
+source.reference.curator-recipes-2.8.0.jar=release/modules/ext/curator-recipes-2.8.0-sources.jar
 spec.version.base=10.7
