+++ Core/src/org/sleuthkit/autopsy/coordinationservice/CoordinationService.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.experimental.coordinationservice;
+package org.sleuthkit.autopsy.coordinationservice;
 
 import java.util.HashMap;
 import java.util.Map;
